dedup: split command into focused modules
scanning, comparison, follow synthesis, and reporting each get a
module, and the auto-dedup that ran inside update moves in with them.
scan targets key on locked revision so revisions of one source scan
independently, transitive scan failures surface as diagnostics instead
of vanishing, and the report and auto-dedup paths share one cached
github compare policy.
